backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
